backend modified

+++ tictactoe/realtime/datamanage.py
@@ -1,5 +1,6 @@
 from channels.db import database_sync_to_async
 from .models import SecondClient
+from asgiref.sync import async_to_sync
 
 
 class DatabaseMangement:
@@ -15,7 +16,6 @@ class DatabaseMangement:
     def create_group(self, group_name, client_name):
         group = SecondClient(group_name=group_name, first_client=client_name)
         group.save()
-        return
 
     @database_sync_to_async
     def create_and_return_first_client(self, group_name, client_name):
@@ -26,4 +26,6 @@ class DatabaseMangement:
 
     @database_sync_to_async
     def delete_group(self, group_name):
-        SecondClient.objects.get(group_name=group_name).delete()
+        print(async_to_sync(self.exist_group)(group_name))
+        if SecondClient.objects.filter(group_name=group_name).first():
+            SecondClient.objects.get(group_name=group_name).delete()
